ironman_stats/requirements.txt
Django==1.8.5
appnope==0.1.0
beautifulsoup4==4.4.1
decorator==4.0.4
django-extensions==1.5.7
django-model-utils==2.3.1
gnureadline==6.3.3
ipdb==0.8.1
ipython==4.0.0
ipython-genutils==0.1.0
lxml==3.4.4
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.5
psycopg2==2.6.1
ptyprocess==0.5
selenium==2.48.0
simplegeneric==0.8.1
six==1.10.0
traitlets==4.0.0
